refactor: member, accommodation dto 패키지 구조 수정 - 도메인에서 dto를 가지고 있도록 구현
